Move everything into singlelinenames, tidy up
